+++ packages/gatsby/cache-dir/__tests__/find-path.js
@@ -18,6 +18,16 @@ describe(`find-path`, () => {
       global.__BASE_PATH__ = `/blog`
       expect(cleanPath(`/blog/mypath`)).toBe(`/mypath`)
     })
+
+    it(`strip out a complex basePrefix`, () => {
+      global.__BASE_PATH__ = `/test/blog`
+      expect(cleanPath(`/test/blog/new`)).toBe(`/new`)
+    })
+
+    it(`strip out an encoded basePrefix`, () => {
+      global.__BASE_PATH__ = encodeURIComponent(`/тест`)
+      expect(cleanPath(`/тест/mypath`)).toBe(`/mypath`)
+    })
   })
 
   describe(`findMatchPath`, () => {