+++ packages/gatsby/src/schema/__tests__/rebuild-schema.js
@@ -479,6 +479,9 @@ describe(`build and update individual types`, () => {
         nodes: [Bar!]!
         pageInfo: PageInfo!
         distinct(field: BarFieldsEnum!): [String!]!
+        max(field: BarFieldsEnum!): Float
+        min(field: BarFieldsEnum!): Float
+        sum(field: BarFieldsEnum!): Float
         group(skip: Int, limit: Int, field: BarFieldsEnum!): [BarGroupConnection!]!
       }"
     `)